+++ WaveFunctionCollapse/IWFCRule.h
@@ -6,13 +6,13 @@
 class IWFCRule {
 protected:
 	std::vector<WFCPosition*> localPositionsWeCareAbout;
-	unsigned long goal;
-	unsigned long tile;
+	unsigned long long goal;
+	unsigned long long tile;
 public:
-	IWFCRule(unsigned long goal, unsigned long tile, std::vector<WFCPosition*> localPositionsWeCareAbout);
+	IWFCRule(unsigned long long goal, unsigned long long tile, std::vector<WFCPosition*> localPositionsWeCareAbout);
 	virtual bool Test(WFCCellUpdate update, const WFCPosition* currentCellPosition) = 0;
-	unsigned long const GetGoal();
-	unsigned long const GetTile();
+	unsigned long long const GetGoal();
+	unsigned long long const GetTile();
 	std::vector< WFCPosition*> GetPositions();
 };
