+++ WaveFunctionCollapse/WFCCellUpdate.cpp
@@ -1,7 +1,7 @@
 #include "WFCCellUpdate.h"
 #include "WFCCell.h"
 
-WFCCellUpdate::WFCCellUpdate(unsigned long removed, unsigned long added, unsigned long collapsed, const WFCPosition* position):
+WFCCellUpdate::WFCCellUpdate(unsigned long long removed, unsigned long long added, unsigned long long collapsed, const WFCPosition* position):
 	updatedCell(position)
 {
 	removedFromDomain = removed;
