+++ WaveFunctionCollapse/WFCRuleCellIsNot.h
@@ -4,7 +4,7 @@ class WFCRuleCellIsNot :
     public IWFCRule
 {
 public:
-    WFCRuleCellIsNot(unsigned long goal, unsigned long tile, std::vector<WFCPosition*> localPositionsWeCareAbout);
+    WFCRuleCellIsNot(unsigned long long goal, unsigned long long tile, std::vector<WFCPosition*> localPositionsWeCareAbout);
     virtual bool Test(WFCCellUpdate update, const WFCPosition* currentCellPosition) override;
 };
